requirements.txt
urllib3==1.22
Django
psycopg2-binary
pytz
sqlparse
django-cors-headers
jsonschema
lxml
requests
python-memcached
werkzeug
pycryptodome
pymemcache
gevent
psutil
PyJWT==1.7.1
mysqlclient
bcrypt
